Created the HttpCheckbox Component
