+++ client/src/pages/ViewEvent/VolunteersTable.tsx
@@ -16,56 +16,12 @@ import check_icon from "../../assets/checkbox-icon.svg";
 import { ProcessedScheduledSlot } from "../../types";
 import { useAuth } from "../../contexts/AuthContext";
 import { ContactPopup } from "../../components/ContactPopup";
+import { HttpCheckbox } from "../../components/HttpCheckbox";
+
 
 /*
 TODO: There is a lot of stuff going on in this component, and we should perhaps look into refactoring at some point. 
 */
-
-//Used for "Confirmed" and "Not going" checkboxes
-interface HttpCheckboxProps {
-  checked: boolean;
-  mutationFn: any; //This needs to be generic enough
-  onSuccess: () => void;
-  onError: () => void;
-}
-export const HttpCheckbox: React.FC<HttpCheckboxProps> = ({
-  checked,
-  mutationFn,
-  onSuccess,
-  onError,
-}: HttpCheckboxProps) => {
-  const [isChecked, setIsChecked] = useState(checked);
-  const httpRequest = useMutation({
-    mutationFn: mutationFn,
-    onSuccess(data, variables, context) {
-      setIsChecked((prevVal) => !prevVal);
-      onSuccess();
-    },
-    onError(error, variables, context) {
-      console.error(error);
-      onError();
-    },
-  });
-
-  //UI
-  return (
-    <div className="relative flex h-full items-center justify-center">
-      {httpRequest.status === "loading" ? (
-        <Loading size="xsmall" thickness="thin" />
-      ) : (
-        <RadixCheckbox.Root
-          className="flex h-5 w-5 items-center justify-center rounded border-2 border-newLeafGreen bg-softGrayWhite shadow-md hover:brightness-110"
-          checked={isChecked}
-          onClick={() => httpRequest.mutate()}
-        >
-          <RadixCheckbox.Indicator className="CheckboxIndicator">
-            <img src={check_icon} alt="" />
-          </RadixCheckbox.Indicator>
-        </RadixCheckbox.Root>
-      )}
-    </div>
-  );
-};
 
 interface FilterItemProps {
   onSelect: () => void;
